Change float to double and move customer

+++ src/account/Account.java
@@ -5,20 +5,21 @@ import report.Report;
 import java.util.ArrayList;
 import java.util.List;
 
+import customer.Customer;
 import promotion.Promotion;
 
 public abstract class Account {
-	protected float balance;
+	protected double balance;
 	protected Region homeregion;
 	protected Customer customer;
 	//protected List<Promotion> promotions;
 	protected Promotion promotion;
-	public abstract Report generateReport(float spentMinutes, float spentMessages,
-			float usedMobileData, float leftMinutes, float leftMessages,
-			float leftMobileData, float costUntilNow);
+	public abstract Report generateReport(double spentMinutes, double spentMessages,
+			double usedMobileData, double leftMinutes, double leftMessages,
+			double leftMobileData, double costUntilNow);
 	public abstract void computeDiscount();
 	
-	public Account(float balance, Region homeregion,Promotion promotion,Customer customer) {
+	public Account(double balance, Region homeregion,Promotion promotion,Customer customer) {
 		super();
 		this.customer = customer;
 		this.balance = balance;
