+++ src/promotion/FriendPromotion.java
@@ -1,4 +1,5 @@
 package promotion;
+import customer.Customer;
 import account.*;
 import report.Report;
 public class FriendPromotion extends Promotion {
@@ -19,7 +20,9 @@ public class FriendPromotion extends Promotion {
 		
 	}
 	
-	public Report generateReport(float spentMinutes, float spentMessages,float usedMobileData, float leftMinutes, float leftMessages,float leftMobileData, float costUntilNow)
+	public Report generateReport(double spentMinutes, double spentMessages,
+			double usedMobileData, double leftMinutes, double leftMessages,
+			double leftMobileData, double costUntilNow)
 	{
 		Report r = super.generateReport(spentMinutes,spentMessages,usedMobileData,leftMinutes,leftMessages,leftMobileData,costUntilNow);
 		r.setDifferentProviderMinutes(differentProviderMinutes);
